Library to contain geometry literals not Promise<> types
